# [Code wide] Fix pending todo + schema + several small improvements