--- a/docs/src/tutorials/heart_rate_analysis.ipynb
+++ b/docs/src/tutorials/heart_rate_analysis.ipynb
@@ -1801,9 +1801,9 @@
     "Currently, heart rate and heart rate variability are calculated at every heartbeat, resulting in a large amount of non-uniform data. To enhance compatibility with other systems, a better approach would be to compute the heart rate at exact 10-second intervals instead. This approach will simplify the data analysis process and provide more consistent and uniform results.\n",
     "\n",
     "The `tp.tick` function generates a uniform timestamp.\n",
-    "The `tp.sample` function sample a signal according to a secondary sampling.\n",
+    "The `tp.resample` function sample a signal according to a secondary sampling.\n",
     "\n",
-    "**Note:** Window functions like tp.`moving_count` have an optional `sampling` argument which is equivalent, but currently more efficient than calling `tp.moving_count` and `tp.sample` separately. Our dataset is small enought that it does not matter."
+    "**Note:** Window functions like tp.`moving_count` have an optional `sampling` argument which is equivalent, but currently more efficient than calling `tp.moving_count` and `tp.resample` separately. Our dataset is small enought that it does not matter."
    ]
   },
   {
@@ -2153,17 +2153,17 @@
     "uniform_sampling_10s = uniform_sampling_10s // 1000\n",
     "uniform_sampling_10s = tp.equal(\n",
     "    uniform_sampling_10s,\n",
-    "    tp.sample(tp.lag(uniform_sampling_10s, 0.01), uniform_sampling_10s) + 1,\n",
+    "    tp.resample(tp.lag(uniform_sampling_10s, 0.01), uniform_sampling_10s) + 1,\n",
     ")\n",
     "uniform_sampling_10s = tp.filter(\n",
     "    tp.select(uniform_sampling_10s, []), uniform_sampling_10s\n",
     ")\n",
     "\n",
     "heart_rate_every_10s = tp.rename(\n",
-    "    tp.sample(heart_rate, sampling=uniform_sampling_10s), \"heart rate (10s)\"\n",
+    "    tp.resample(heart_rate, sampling=uniform_sampling_10s), \"heart rate (10s)\"\n",
     ")\n",
     "heart_rate_variability_every_10s = tp.rename(\n",
-    "    tp.sample(heart_rate_variability, sampling=uniform_sampling_10s),\n",
+    "    tp.resample(heart_rate_variability, sampling=uniform_sampling_10s),\n",
     "    \"heart rate variability (10s)\",\n",
     ")\n",
     "\n",