--- a/docs/src/tutorials/loan_outcomes_prediction.ipynb
+++ b/docs/src/tutorials/loan_outcomes_prediction.ipynb
@@ -588,10 +588,15 @@
     "df_trx[\"date\"] = pd.to_datetime(df_trx[\"date\"])\n",
     "\n",
     "# Remove unused columns\n",
-    "df_loan.drop(columns=[\"loan_id\"], inplace=True, errors='ignore')\n",
+    "df_loan.drop(columns=[\"loan_id\"], inplace=True, errors=\"ignore\")\n",
     "\n",
     "# Remove trx without loan and get loan dates into transactions\n",
-    "df_trx = df_trx.join(df_loan[[\"account_id\", \"date\"]], on=['account_id'], how=\"inner\", rsuffix=\"_loan\")\n",
+    "df_trx = df_trx.join(\n",
+    "    df_loan[[\"account_id\", \"date\"]],\n",
+    "    on=[\"account_id\"],\n",
+    "    how=\"inner\",\n",
+    "    rsuffix=\"_loan\",\n",
+    ")\n",
     "\n",
     "# Only transactions before the loan can be used for prediction\n",
     "df_trx[\"valid\"] = df_trx[\"date_loan\"] >= df_trx[\"date\"]\n",
@@ -711,7 +716,9 @@
    ],
    "source": [
     "# Create events from the dataframes\n",
-    "e_loan_data = tp.EventSet.from_dataframe(df_loan, timestamp_column=\"date\", index_names=[\"account_id\"])\n",
+    "e_loan_data = tp.EventSet.from_dataframe(\n",
+    "    df_loan, timestamp_column=\"date\", index_names=[\"account_id\"]\n",
+    ")\n",
     "e_loan_data"
    ]
   },
@@ -802,7 +809,9 @@
     }
    ],
    "source": [
-    "e_trx_data = tp.EventSet.from_dataframe(df_trx, timestamp_column=\"date\", index_names=[\"account_id\"])\n",
+    "e_trx_data = tp.EventSet.from_dataframe(\n",
+    "    df_trx, timestamp_column=\"date\", index_names=[\"account_id\"]\n",
+    ")\n",
     "e_trx_data"
    ]
   },
@@ -883,9 +892,13 @@
     "# Get amount with sign +/- depending on type of transaction\n",
     "# - PRIJEM: deposit\n",
     "# - VYDAJ or VYBER: withdrawal\n",
-    "is_withdrawal = tp.equal_scalar(valid_trx[\"type\"], \"VYDAJ\") + tp.equal_scalar(valid_trx[\"type\"], \"VYBER\")\n",
+    "is_withdrawal = tp.equal_scalar(valid_trx[\"type\"], \"VYDAJ\") + tp.equal_scalar(\n",
+    "    valid_trx[\"type\"], \"VYBER\"\n",
+    ")\n",
     "is_deposit = tp.equal_scalar(valid_trx[\"type\"], \"PRIJEM\")\n",
-    "trx_sign = -1.0 * tp.cast(is_withdrawal, dtype.DType.FLOAT32) + tp.cast((is_deposit), dtype.DType.FLOAT32)\n",
+    "trx_sign = -1.0 * tp.cast(is_withdrawal, dtype.DType.FLOAT32) + tp.cast(\n",
+    "    (is_deposit), dtype.DType.FLOAT32\n",
+    ")\n",
     "trx_signed_amount = trx_sign * tp.cast(valid_trx[\"amount\"], dtype.DType.FLOAT32)\n",
     "\n",
     "# Accumulated amount (signed) should be equal to \"balance\"\n",
@@ -900,8 +913,17 @@
     ")\n",
     "\n",
     "# Rename output features\n",
-    "new_names = [\"amount\", \"balance_registered\", \"balance_calculated\", \"daily_count\", \"monthly_count\"]\n",
-    "result = tp.rename(result, {old_name: new_names[i] for i, old_name in enumerate(result.feature_names)}) \n",
+    "new_names = [\n",
+    "    \"amount\",\n",
+    "    \"balance_registered\",\n",
+    "    \"balance_calculated\",\n",
+    "    \"daily_count\",\n",
+    "    \"monthly_count\",\n",
+    "]\n",
+    "result = tp.rename(\n",
+    "    result,\n",
+    "    {old_name: new_names[i] for i, old_name in enumerate(result.feature_names)},\n",
+    ")\n",
     "\n",
     "# Execute graph\n",
     "result_data = tp.evaluate(result, {source_trx: e_trx_data})"
@@ -943,8 +965,7 @@
     "# Also, use a line instead of default markers (due to non-uniform sampling)\n",
     "\n",
     "result_data.plot(\n",
-    "    features=[\"balance_registered\", \"balance_calculated\"],\n",
-    "    style=\"line\"\n",
+    "    features=[\"balance_registered\", \"balance_calculated\"], style=\"line\"\n",
     ")"
    ]
   },
@@ -1333,7 +1354,11 @@
     }
    ],
    "source": [
-    "result_data.plot(indexes=[(1,), (2,)], features=[\"balance\", \"daily_count\", \"monthly_count\"], interactive=True)"
+    "result_data.plot(\n",
+    "    indexes=[(1,), (2,)],\n",
+    "    features=[\"balance\", \"daily_count\", \"monthly_count\"],\n",
+    "    interactive=True,\n",
+    ")"
    ]
   },
   {