--- a/docs/src/tutorials/toy_example.ipynb
+++ b/docs/src/tutorials/toy_example.ipynb
@@ -260,7 +260,7 @@
     "detrend = signal - trend\n",
     "trend2 = tp.simple_moving_average(detrend, 10)\n",
     "detrend2 = detrend - trend2\n",
-    "slope = trend2 - tp.sample(tp.lag(trend2, 5), trend2)\n",
+    "slope = trend2 - tp.resample(tp.lag(trend2, 5), trend2)\n",
     "\n",
     "# TODO: Clean feature names with rename.\n",
     "result = tp.glue(\n",