--- a/temporian/__init__.py
+++ b/temporian/__init__.py
@@ -20,15 +20,17 @@ from temporian.core import graph
 from temporian.core import serialize
 from temporian.core.data import dtype
 from temporian.core.data import node
-from temporian.core.data import feature
-from temporian.core.data import sampling
+from temporian.core.data import schema
 from temporian.core.data import duration
 from temporian.core.operators import base
 from temporian.io.read_event_set import read_event_set
 from temporian.io.save_event_set import save_event_set
 
 from temporian.implementation.numpy.data.event_set import EventSet
-from temporian.implementation.numpy.data.io import event_set
+from temporian.implementation.numpy.data.io import (
+    event_set,
+    pd_dataframe_to_event_set,
+)
 
 from temporian.implementation.numpy.data.plotter import plot
 
@@ -51,7 +53,6 @@ str_ = dtype.DType.STRING
 __version__ = "0.0.1"
 
 evaluate = evaluation.evaluate
-Feature = feature.Feature
 load = serialize.load
 save = serialize.save
 input_node = node.input_node