--- a/temporian/core/data/duration.py
+++ b/temporian/core/data/duration.py
@@ -20,7 +20,7 @@ This datatype is equivalent to a double in C.
 """
 import datetime
 from enum import Enum
-from typing import Union
+from typing import Union, Optional
 
 import numpy as np
 
@@ -29,6 +29,7 @@ NormalizedDuration = float
 Duration = Union[float, int]
 
 Timestamp = Union[np.datetime64, datetime.datetime, int, float]
+NormalizedTimestamp = float
 
 
 def normalize_duration(x: Duration) -> NormalizedDuration:
@@ -37,8 +38,30 @@ def normalize_duration(x: Duration) -> NormalizedDuration:
 
     raise ValueError(
         "A duration should be a strictly positive number of type float,"
-        f" int or tp.Duration. Got {x!r} or type {type(x)}."
+        f" int or tp.Duration. Got {x!r} of type {type(x)}."
     )
+
+
+def normalize_timestamps(x: Timestamp) -> NormalizedTimestamp:
+    if isinstance(x, np.datetime64):
+        return x.astype("datetime64[ns]").astype(np.float64) / 1e9
+
+    if isinstance(x, datetime.datetime):
+        return float(x.timestamp())
+
+    if isinstance(x, int):
+        return float(x)
+
+    if isinstance(x, float):
+        return x
+
+    raise ValueError(f"Invalid timestamp {x!r} of type {type(x)}.")
+
+
+def normalize_timestamps_or_none(x: Optional[Timestamp]) -> NormalizedTimestamp:
+    if x is None:
+        return None
+    return normalize_timestamps(x)
 
 
 def milliseconds(value: Union[int, float]) -> Duration:
@@ -117,7 +140,7 @@ def weeks(value: Union[int, float]) -> Duration:
     return NormalizedDuration(value * 60 * 60 * 24 * 7)
 
 
-def convert_date_to_duration(date: Timestamp) -> Duration:
+def convert_date_to_duration(date: Timestamp) -> NormalizedDuration:
     """Converts date value to a number representing the Unix timestamp.
 
     If a float or int, it is returned as float.
@@ -152,17 +175,21 @@ def convert_date_to_duration(date: Timestamp) -> Duration:
     raise TypeError(f"Unsupported type: {type(date)}")
 
 
-def convert_numpy_datetime64_to_duration(date: np.datetime64) -> Duration:
+def convert_numpy_datetime64_to_duration(
+    date: np.datetime64,
+) -> NormalizedDuration:
     """Convert numpy datetime64 to duration epoch UTC"""
-    return date.astype("datetime64[s]").astype("float64")
+    return float(date.astype("datetime64[s]").astype("float64"))
 
 
-def convert_datetime_to_duration(date: datetime.datetime) -> Duration:
+def convert_datetime_to_duration(date: datetime.datetime) -> NormalizedDuration:
     """Convert datetime to duration epoch UTC"""
-    return date.replace(tzinfo=datetime.timezone.utc).timestamp()
+    return float(date.replace(tzinfo=datetime.timezone.utc).timestamp())
 
 
-def convert_datetime_date_to_duration(date: datetime.date) -> Duration:
+def convert_datetime_date_to_duration(
+    date: datetime.date,
+) -> NormalizedDuration:
     """Convert date to duration epoch UTC"""
     return convert_datetime_to_duration(
         datetime.datetime.combine(date, datetime.time(0, 0))