--- a/temporian/core/data/node.py
+++ b/temporian/core/data/node.py
@@ -618,8 +618,8 @@ def input_node(
         is_unix_timestamp: If true, the timestamps are interpreted as unix
           timestamps in seconds.
         same_sampling_as: If set, the created node is guarentied to have the
-          same sampling as "same_sampling_as". In this case, "indexes" and
-          "is_unix_timestamp" should not be provided. Some operators require for
+          same sampling as same_sampling_as`. In this case, `indexes` and
+          `is_unix_timestamp` should not be provided. Some operators require for
           input nodes to have the same sampling.
     """