--- a/temporian/test/user_guide_test.py
+++ b/temporian/test/user_guide_test.py
@@ -1,7 +1,5 @@
 from absl.testing import absltest
 
-import temporian as tp
-
 
 class TFPTest(absltest.TestCase):
     def test_all(self):