… .vscode/settings.json
… .vscode/tasks.json
… CONTRIBUTING.md
… benchmark/benchmark_time.py
… benchmark/scripts/basic.py
… benchmark/scripts/sma.py
… docs/src/user_guide.md
… temporian/BUILD
… temporian/core/data/feature.py
… temporian/core/data/sampling.py
… temporian/core/data/schema.py
… temporian/core/data/test/node_test.py
… temporian/core/evaluation.py
… temporian/core/graph.py
… temporian/core/operators/BUILD
… temporian/core/operators/add_index.py
… temporian/core/operators/all_operators.py
… temporian/core/operators/base.py
… temporian/core/operators/binary/BUILD
… temporian/core/operators/binary/arithmetic.py
… temporian/core/operators/binary/base.py
… temporian/core/operators/binary/logical.py
… temporian/core/operators/binary/relational.py
… temporian/core/operators/calendar/BUILD
… temporian/core/operators/calendar/base.py
… temporian/core/operators/cast.py
… temporian/core/operators/drop_index.py
… temporian/core/operators/filter.py
… temporian/core/operators/glue.py
… temporian/core/operators/lag.py