- Add Event Set constructor method that does not rely on Pandas. Add argument to create multiple Nodes or EventSets with the same sampling.
- Add "source_node()" function event sets.
- Rename "tp.input_node" to "tp.source_node".
- Definition and use of "schema" in the Node and EventSet.
- Add utility functions to handle Nodes and EventSets.
- Update of all ops and tests with new schema system.
- Various code optimizations e.g. removed unnecessary sorts, list creation, dictionary query, shuffle for-loops, etc.
- Various code cleaning e.g. simplify unit tests.
- Various code improvements e.g. more checks, fix annotations, remove redundant code, de-duplicate default arguments, error messages, unit tests utility, solve some memory handling errors, remove private attribute usage, simplify code.
- Remove argument type overloading in all operator / operator implementation constructors.
- set_index now keeps the removed index e.g., convert them into features.
- Rename sample operator to resample.
- Remove @property on all non N(1) time-cost functions e.g. node.schema.feature_names() instead of node.schema.feature_names.
- Normalize data dtypes across all operators / utilities.
- Update pandas dataframe importer to use event set user constructor.
- Support for List[DTYPE] in operator attributes.
- Remove support for list of durations in lag and leak operators.
- Split lag and leak into two separate operators.
- drop_index adds the new feature at the end.
- add_index adds the new index at the end.
- Remove indexing implementation in Pandas importer. Instead, importing EventSet (directy or with a pd dataframe) uses the add_index operator internally.
- The evaluation algorithm does not rely anymore on node's string representation to index the nodes.
- Remove support for named nodes in evaluation. To be thought again.
- Graph object support named and unnamed nodes. The algorithm to handle node names is separated from the graph inference algorithm. Named nodes are only required for serialization (so far).
- Add support for "evaluate" on event set directly i.e. the source node is extracted automatically.


**Benchmark**

Main observations.
- No change in most ops.
- from_dataframe: 17x speed-up
- add_index / set_index: from 1.8x slow-down, to 2x speedup depending on the experiment. Added todo for further improvement in the code.

```
AFTER

================================================================
Name                              Wall time (s)    CPU time (s)
================================================================
from_dataframe:s:10_000_numidx:0_numidxval:20_idx:int       0.00030       0.00030
from_dataframe:s:10_000_numidx:0_numidxval:20_idx:str       0.00028       0.00028
from_dataframe:s:10_000_numidx:1_numidxval:20_idx:int       0.00280       0.00280
from_dataframe:s:10_000_numidx:1_numidxval:20_idx:str       0.00577       0.00578
from_dataframe:s:10_000_numidx:3_numidxval:20_idx:int       0.03032       0.03031
from_dataframe:s:10_000_numidx:3_numidxval:20_idx:str       0.03728       0.03728
from_dataframe:s:10_000_numidx:5_numidxval:20_idx:int       0.04636       0.04634
from_dataframe:s:10_000_numidx:5_numidxval:20_idx:str       0.06261       0.06258
----------------------------------------------------------------
simple_moving_average:100            0.00059       0.00059
simple_moving_average:10_000         0.00091       0.00091
simple_moving_average:1_000_000       0.00597       0.00597
----------------------------------------------------------------
select_and_glue:100                  0.00043       0.00043
select_and_glue:10_000               0.00292       0.00292
select_and_glue:1_000_000            0.00063       0.00063
----------------------------------------------------------------
calendar_day_of_month:100            0.00009       0.00009
calendar_day_of_month:10_000         0.00487       0.00487
calendar_day_of_month:1_000_000       0.50778       0.50776
----------------------------------------------------------------
sample:e100_s100                     0.00065       0.00065
sample:e100_s10_000                  0.00282       0.00282
sample:e100_s1_000_000               0.00356       0.00356
sample:e10_000_s100                  0.00065       0.00065
sample:e10_000_s10_000               0.00096       0.00096
sample:e10_000_s1_000_000            0.00460       0.00460
sample:e1_000_000_s100               0.00086       0.00086
sample:e1_000_000_s10_000            0.00137       0.00137
sample:e1_000_000_s1_000_000         0.00685       0.00685
----------------------------------------------------------------
propagate:100                        0.00034       0.00034
propagate:10_000                     0.00048       0.00048
propagate:1_000_000                  0.00048       0.00048
----------------------------------------------------------------
cast(check=False):100                0.00024       0.00024
cast(check=True):100                 0.00091       0.00091
cast(check=False):1000000            0.00163       0.00163
cast(check=True):1000000             0.00315       0.00315
----------------------------------------------------------------
unique_timestamps:100                0.00047       0.00047
unique_timestamps:10000              0.00072       0.00072
unique_timestamps:1000000            0.00672       0.00672
----------------------------------------------------------------
add_index:s:10_000:num_idx:1         0.00447       0.00447
add_index:s:10_000:num_idx:2         0.01682       0.01682
add_index:s:10_000:num_idx:3         0.04388       0.04388
add_index:s:10_000:num_idx:4         0.04627       0.04627
add_index:s:10_000:num_idx:5         0.04057       0.04056
add_index:s:100_000:num_idx:1        0.03086       0.03085
add_index:s:100_000:num_idx:2        0.04657       0.04657
add_index:s:100_000:num_idx:3        0.17663       0.17663
add_index:s:100_000:num_idx:4        0.47043       0.47041
add_index:s:100_000:num_idx:5        0.46969       0.46964
add_index:s:1_000_000:num_idx:1       0.30103       0.30101
add_index:s:1_000_000:num_idx:2       0.34583       0.34582
add_index:s:1_000_000:num_idx:3       0.51490       0.51486
add_index:s:1_000_000:num_idx:4       1.89817       1.89801
add_index:s:1_000_000:num_idx:5       4.92727       4.92624
================================================================

BEFORE

================================================================
Name                              Wall time (s)    CPU time (s)
================================================================
from_dataframe:s:10_000_numidx:0_numidxval:20_idx:int       0.00041       0.00041
from_dataframe:s:10_000_numidx:0_numidxval:20_idx:str       0.00040       0.00040
from_dataframe:s:10_000_numidx:1_numidxval:20_idx:int       0.00188       0.00188
from_dataframe:s:10_000_numidx:1_numidxval:20_idx:str       0.00489       0.00489
from_dataframe:s:10_000_numidx:3_numidxval:20_idx:int       0.33466       0.33460
from_dataframe:s:10_000_numidx:3_numidxval:20_idx:str       0.54433       0.54430
from_dataframe:s:10_000_numidx:5_numidxval:20_idx:int       0.59611       0.59608
from_dataframe:s:10_000_numidx:5_numidxval:20_idx:str       1.11714       1.11645
----------------------------------------------------------------
simple_moving_average:100            0.00055       0.00055
simple_moving_average:10_000         0.00085       0.00085
simple_moving_average:1_000_000       0.00752       0.00752
----------------------------------------------------------------
select_and_glue:100                  0.00034       0.00034
select_and_glue:10_000               0.00047       0.00047
select_and_glue:1_000_000            0.00045       0.00045
----------------------------------------------------------------
calendar_day_of_month:100            0.00008       0.00008
calendar_day_of_month:10_000         0.00479       0.00479
calendar_day_of_month:1_000_000       0.47936       0.47933
----------------------------------------------------------------
sample:e100_s100                     0.00061       0.00061
sample:e100_s10_000                  0.00079       0.00079
sample:e100_s1_000_000               0.00351       0.00351
sample:e10_000_s100                  0.00059       0.00059
sample:e10_000_s10_000               0.00095       0.00095
sample:e10_000_s1_000_000            0.00469       0.00469
sample:e1_000_000_s100               0.00087       0.00087
sample:e1_000_000_s10_000            0.00140       0.00140
sample:e1_000_000_s1_000_000         0.00718       0.00718
----------------------------------------------------------------
propagate:100                        0.00033       0.00033
propagate:10_000                     0.00047       0.00047
propagate:1_000_000                  0.00045       0.00045
----------------------------------------------------------------
cast(check=False):100                0.00021       0.00021
cast(check=True):100                 0.00100       0.00099
cast(check=False):1000000            0.00153       0.00153
cast(check=True):1000000             0.00325       0.00325
----------------------------------------------------------------
unique_timestamps:100                0.00052       0.00051
unique_timestamps:10000              0.00077       0.00077
unique_timestamps:1000000            0.00725       0.00725
----------------------------------------------------------------
set_index:s:10_000:num_idx:1:append:False       0.01329       0.01329
set_index:s:10_000:num_idx:2:append:False       0.01483       0.01483
set_index:s:10_000:num_idx:3:append:False       0.02704       0.02704
set_index:s:10_000:num_idx:4:append:False       0.07546       0.07546
set_index:s:10_000:num_idx:5:append:False       0.07996       0.07995
set_index:s:100_000:num_idx:1:append:False       0.16087       0.16086
set_index:s:100_000:num_idx:2:append:False       0.16262       0.16262
set_index:s:100_000:num_idx:3:append:False       0.15592       0.15592
set_index:s:100_000:num_idx:4:append:False       0.24693       0.24692
set_index:s:100_000:num_idx:5:append:False       0.93581       0.93576
set_index:s:1_000_000:num_idx:1:append:False       2.38795       2.38752
set_index:s:1_000_000:num_idx:2:append:False       2.45258       2.44883
set_index:s:1_000_000:num_idx:3:append:False       1.52917       1.52897
set_index:s:1_000_000:num_idx:4:append:False       1.54799       1.54789
set_index:s:1_000_000:num_idx:5:append:False       2.73235       2.73212
================================================================
```