@achoum 2023-05-31
Thanks for the initial review. Let me address and solve the nits.

For the larger questions, I've added some material to our next meeting agenda. We can go over it and validate / invalidate those changes.