@ianspektor 2023-06-01
One last comment for thoroughness on docstring formatting: running the docs (`mkdocs serve -f docs/mkdocs.yml`) prints out a list of indentation errors and parameter description mismatches :) (and these latter should definitely be fixed)

```
WARNING  -  griffe: temporian/core/operators/propagate.py:120: Confusing indentation for continuation line 27 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/core/data/node.py:522: Confusing indentation for continuation line 23 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/core/data/node.py:524: Confusing indentation for continuation line 25 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/core/data/node.py:526: Confusing indentation for continuation line 27 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/core/data/node.py:527: Confusing indentation for continuation line 28 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/core/data/node.py:528: Confusing indentation for continuation line 29 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:49: Confusing indentation for continuation line 21 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:52: Confusing indentation for continuation line 24 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:53: Confusing indentation for continuation line 25 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:54: Confusing indentation for continuation line 26 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:56: Confusing indentation for continuation line 28 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:57: Confusing indentation for continuation line 29 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:166: Confusing indentation for continuation line 15 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:162: No type or annotation for parameter 'is_sorted'
WARNING  -  griffe: temporian/implementation/numpy/data/io.py:162: Parameter 'is_sorted' does not appear in the function signature
WARNING  -  griffe: temporian/implementation/numpy/data/event_set.py:392: Confusing indentation for continuation line 6 in docstring, should be 4 * 2 = 8 spaces, not 6
WARNING  -  griffe: temporian/implementation/numpy/data/event_set.py:391: No type or annotation for parameter 'force_new_node'
```