@ianspektor 2023-06-01
And a similar one for unused imports with `flake8 --select F401 --per-file-ignores="__init__.py:F401 all_operators.py:F401" temporian`:
```
temporian/core/test/registered_operators_test.py:20:1: F401 'temporian.core.operators.all_operators as _op' imported but unused
temporian/core/data/node.py:28:5: F401 'temporian.implementation.numpy.data.event_set.EventSet' imported but unused
temporian/core/data/schema.py:25:5: F401 'temporian.core.operators.base.Operator' imported but unused
temporian/test/user_guide_test.py:3:1: F401 'temporian as tp' imported but unused
temporian/implementation/numpy/evaluation.py:26:1: F401 'temporian.implementation.numpy.operators.all_operators as _impls' imported but unused
temporian/implementation/numpy/test/registered_operators_test.py:18:1: F401 'temporian.implementation.numpy.operators.all_operators as imps' imported but unused
temporian/implementation/numpy/operators/test/since_last_test.py:21:1: F401 'temporian.implementation.numpy.operators.since_last.operators_cc' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:25:5: F401 'bokeh.plotting.figure' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:27:5: F401 'bokeh.layouts.column' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:28:5: F401 'bokeh.models.ColumnDataSource' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:28:5: F401 'bokeh.models.CategoricalColorMapper' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:28:5: F401 'bokeh.models.HoverTool' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:196:5: F401 'bokeh.io.output_notebook' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:196:5: F401 'bokeh.io.show' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:197:5: F401 'bokeh.layouts.gridplot' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:197:5: F401 'bokeh.layouts.column' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:198:5: F401 'bokeh.models.CategoricalColorMapper' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:198:5: F401 'bokeh.models.HoverTool' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:199:5: F401 'bokeh.models.ColumnDataSource' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:199:5: F401 'bokeh.models.CustomJS' imported but unused
temporian/implementation/numpy/data/plotter_bokeh.py:200:5: F401 'bokeh.palettes.Dark2_5 as colors' imported but unused
temporian/implementation/numpy/data/test/plotter_test.py:18:13: F401 'IPython.display' imported but unused
```